move existing comp/tooltip component
